Test to make sure value is a string before trying to JSON.parse()
I'm using PG 10.1, the latest AdonisJS version, and a `jsonb` column. When I use this trait and query for a record with a field in the `jsonField` list that is holding an object, the value coming out of Lucid is actually an object and doesn't need to be `JSON.parse()`d.  It's actually the same for saving an object as well.  But you do need this trait for arrays (i.e. `['a', 'b', 'c']`).  So this will do a tiny bit of checking to make sure what is being parse is a string and not already an object.